move all todo logic into context file
