+++ src/TodosContext.tsx
@@ -1,16 +1,5 @@
-import React, { createContext, useState } from 'react';
-
-type Todo = {
-  id: number;
-  title: string;
-  completed: boolean;
-};
-
-type TodosContextType = {
-  todos: Todo[];
-  addTodo: (title: string) => void;
-  toggleTodo: (id: number) => void;
-};
+import React, { ChangeEvent, createContext, useContext, useState } from "react";
+import { TodosContextType, Todo } from "./types";
 
 const TodosContext = createContext<TodosContextType | undefined>(undefined);
 
@@ -18,6 +7,7 @@ type Props = { children: React.ReactNode };
 
 export const TodosProvider: React.FC<Props> = ({ children }) => {
   const [todos, setTodos] = useState<Todo[]>([]);
+  const [title, setTitle] = useState("");
 
   const addTodo = (title: string) => {
     const newTodo = {
@@ -27,17 +17,46 @@ export const TodosProvider: React.FC<Props> = ({ children }) => {
     };
 
     setTodos([...todos, newTodo]);
+    setTitle("");
+  };
+
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    addTodo(title);
   };
 
   const toggleTodo = (id: number) => {
-    const updatedTodos = todos.map((todo) => (todo.id === id ? { ...todo, completed: !todo.completed } : todo));
+    const updatedTodos = todos.map((todo) =>
+      todo.id === id ? { ...todo, completed: !todo.completed } : todo
+    );
 
     setTodos(updatedTodos);
   };
 
+  const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
+    setTitle(e.target.value);
+  };
+
   return (
-    <TodosContext.Provider value={{ todos, addTodo, toggleTodo }}>
+    <TodosContext.Provider
+      value={{
+        todos,
+        addTodo,
+        toggleTodo,
+        handleSubmit,
+        handleInputChange,
+        title,
+      }}
+    >
       {children}
     </TodosContext.Provider>
   );
+};
+
+export const useTodosContext = () => {
+  const context = useContext(TodosContext);
+  if (context === undefined) {
+    throw new Error("useTodosContext must be used with a context");
+  }
+  return context;
 };
